+++ returns/future.py
@@ -17,7 +17,8 @@ from typing_extensions import final
 
 from returns._generated.futures import _future, _future_result
 from returns._generated.iterable import iterable
-from returns.interfaces import applicative, bindable, mappable
+from returns.interfaces import applicative, bindable, mappable, rescuable
+from returns.interfaces.specific import result
 from returns.io import IO, IOResult
 from returns.primitives.container import BaseContainer
 from returns.primitives.hkt import Kind1, Kind2, dekind
@@ -495,6 +496,8 @@ class FutureResult(
     mappable.Mappable2[_ValueType, _ErrorType],
     bindable.Bindable2[_ValueType, _ErrorType],
     applicative.Applicative2[_ValueType, _ErrorType],
+    rescuable.Rescuable2[_ValueType, _ErrorType],
+    result.ResultBased2[_ValueType, _ErrorType],
 ):
     """
     Container to easily compose ``async`` functions.
@@ -1023,7 +1026,7 @@ class FutureResult(
         self,
         function: Callable[
             [_ErrorType],
-            'FutureResult[_ValueType, _NewErrorType]',
+            Kind2['FutureResult', _ValueType, _NewErrorType],
         ],
     ) -> 'FutureResult[_ValueType, _NewErrorType]':
         """