+++ returns/methods/__init__.py
@@ -1,3 +1,5 @@
 from returns.methods.apply import apply as apply
 from returns.methods.bind import bind as bind
+from returns.methods.bind_result import bind_result as bind_result
 from returns.methods.map import map_ as map_
+from returns.methods.rescue import rescue as rescue